+++ CPUPipeline/src/Shader.cpp
@@ -60,7 +60,7 @@ std::string Shader::ReadFile(const std::string& filepath)
 	if (in)
 	{
 		in.seekg(0, std::ios::end);
-		size_t size = in.tellg();
+		size_t size = (size_t)in.tellg();
 		if (size != -1)
 		{
 			result.resize(size);
